create reward-phase.js and added relative rewardPhase functions
