+++ utils/eventHandler.js
@@ -1,32 +1,32 @@
-import { config } from '../take.js';
-import { player } from '../utils/config.js';
+import { config } from "../take.js";
+import { player } from "../utils/config.js";
 import {
   indexOfHighlightBox,
   indexOfHighlightBoxModify,
   numberOfReward,
-} from './utility.js';
+} from "./Reward-phase.js";
 
 function playerKeydownHandler(e) {
   if (config.inGamePhase) {
     switch (e.key) {
-      case 'w':
-      case 'ArrowUp': {
+      case "w":
+      case "ArrowUp": {
         player.moveUp = true;
         break;
       }
-      case 's':
-      case 'ArrowDown': {
+      case "s":
+      case "ArrowDown": {
         player.moveDown = true;
 
         break;
       }
-      case 'a':
-      case 'ArrowLeft': {
+      case "a":
+      case "ArrowLeft": {
         player.moveLeft = true;
         break;
       }
-      case 'd':
-      case 'ArrowRight': {
+      case "d":
+      case "ArrowRight": {
         player.moveRight = true;
         break;
       }
@@ -36,15 +36,15 @@ function playerKeydownHandler(e) {
   }
   if (config.inRewardPhase) {
     switch (e.key) {
-      case 'w':
-      case 'ArrowUp': {
+      case "w":
+      case "ArrowUp": {
         if (indexOfHighlightBox > 0) {
           indexOfHighlightBoxModify(indexOfHighlightBox - 1);
         }
         break;
       }
-      case 's':
-      case 'ArrowDown': {
+      case "s":
+      case "ArrowDown": {
         if (indexOfHighlightBox < numberOfReward - 1) {
           indexOfHighlightBoxModify(indexOfHighlightBox + 1);
           break;
@@ -59,24 +59,24 @@ function playerKeydownHandler(e) {
 function playerKeyupHandler(e) {
   if (config.inGamePhase) {
     switch (e.key) {
-      case 'w':
-      case 'ArrowUp': {
+      case "w":
+      case "ArrowUp": {
         player.moveUp = false;
         break;
       }
-      case 's':
-      case 'ArrowDown': {
+      case "s":
+      case "ArrowDown": {
         player.moveDown = false;
 
         break;
       }
-      case 'a':
-      case 'ArrowLeft': {
+      case "a":
+      case "ArrowLeft": {
         player.moveLeft = false;
         break;
       }
-      case 'd':
-      case 'ArrowRight': {
+      case "d":
+      case "ArrowRight": {
         player.moveRight = false;
         break;
       }
